map: update styling for seasons
Each season is displayed with a color similar to the legend at the
right. Uses CartoCSS to do this.

A few experiments were run to limit the number of points to the map
with some gis queries but they were too expensive to run on demand.
Might be worth a different dataset that has totals within a certain
resolution (100 meters or someting). The experiments are commented
out in the crash-map.js.